Fix strict warnings and compiler errors across the codebase.
- Ensure proper C_CDD_EXPORT usage in cdd_c_error.h.
- Fix array initialization using proper const char* or casting to prevent -Wcast-qual drops.
- Adjust -Wno-overlength-strings usage.
- Fix C99-style declaration-after-statement compiler errors.
- Change AZ_SPAN_FROM_STR to az_span_create_from_str where needed to prevent string length compilation errors on casted literals.

+++ src/tests/parse/test_cdd_cst_semantic.h
@@ -1151,7 +1151,8 @@ TEST test_cdd_cst_parser_oom_new(void) {
 
     g_cdd_alloc_fail = i;
     {
-      int rc = cdd_cst_parse(az_span_create_from_str((char *)code), &tree);
+      int rc =
+          cdd_cst_parse(az_span_create_from_str((char *)(size_t)code), &tree);
       g_cdd_alloc_fail = 0;
 
       if (rc == 0) {
